[KAN-90] refactor: useParticipation 훅 생성 후 분리
