+++ src/app/components/BottomFloatingBar/ParticipationButton.tsx
@@ -1,17 +1,15 @@
 'use client';
 
-import { useEffect, useState } from 'react';
+import { useEffect } from 'react';
 import { useParams, useRouter } from 'next/navigation';
 
 import { UserData } from '@/types/client.type';
 import Button from '../Button/Button';
 import useCopyUrlToClipboard from '@/hooks/useCopyUrlToClipboard';
 import useCancelGathering from '@/hooks/useCancelGathering';
-import postGatheringToJoin from '@/app/api/actions/gatherings/postGatheringToJoin';
-import deleteGatheringToWithdraw from '@/app/api/actions/gatherings/deleteGatheringToWithdraw';
+import useParticipation from '@/hooks/useParticipation';
 import { GatheringParticipantsType } from '@/types/data.type';
 import Popup from '../Popup/Popup';
-import toast from 'react-hot-toast';
 
 interface ParticipationButtonProps {
   isHost: boolean;
@@ -38,8 +36,14 @@ const ParticipationButton = ({
   const { copyUrlToClipboard } = useCopyUrlToClipboard();
   const { cancelGathering } = useCancelGathering(Number(params.id));
 
-  const [hasParticipated, setHasParticipated] = useState<boolean>(false);
-  const [showPopup, setShowPopup] = useState<boolean>(false);
+  const {
+    hasParticipated,
+    setHasParticipated,
+    showPopup,
+    setShowPopup,
+    handleJoinClick,
+    handleWithdrawClick,
+  } = useParticipation(user);
 
   useEffect(() => {
     if (user) {
@@ -48,6 +52,7 @@ const ParticipationButton = ({
       );
       setHasParticipated(isParticipated);
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [user, participantsData]);
 
   const isFull = participantCount === capacity; //참여인원이 가득찼는지 검사
@@ -61,55 +66,6 @@ const ParticipationButton = ({
     : isParticipationDisabled
       ? 'gray'
       : 'default';
-
-  const handleJoinClick = async () => {
-    if (!user) {
-      // 유저가 없는 경우
-      setShowPopup(true); // 팝업 표시
-      return;
-    }
-
-    if (!hasParticipated) {
-      const { success, message } = await postGatheringToJoin(Number(params.id));
-
-      // 성공하지 못한 경우
-      if (!success) {
-        toast.error(message, {
-          className: 'text-14 font-bold',
-        });
-
-        return;
-      }
-
-      toast.success(message, {
-        className: 'text-14 font-bold',
-      });
-
-      setHasParticipated(true);
-    }
-  };
-
-  const handleWithdrawClick = async () => {
-    if (hasParticipated) {
-      const { success, message } = await deleteGatheringToWithdraw(
-        Number(params.id),
-      );
-
-      if (!success) {
-        toast.error(message, {
-          className: 'text-14 font-bold',
-        });
-
-        return;
-      }
-
-      toast.success(message, {
-        className: 'text-14 font-bold',
-      });
-
-      setHasParticipated(false);
-    }
-  };
 
   /* 버튼 렌더링 함수 */
   const renderButton = (
